+++ src/RcppDistance.cpp
@@ -85,25 +85,33 @@ bool seqEqual(std::string seq1, std::string seq2,
 //' \code{pairwiseEqual} determined pairwise equivalence between a pairs in a 
 //' set of sequences, excluding ambiguous positions (Ns and gaps).
 //'
-//' @param    seq  character vector containing a DNA sequences.
+//' @param    seq     character vector containing a DNA sequences.
+//' @param    ignore  vector of characters to ignore when testing for equality.
+//'                   Default is to ignore c("N","-",".","?")
 //'
-//' @return   A logical matrix of equivalence between each entry in \code{seq}. 
+//' @return   A logical matrix of equivalence between each entry in \code{seq}.
 //'           Values are \code{TRUE} when sequences are equivalent and \code{FALSE}
 //'           when they are not.
-//' 
+//'
 //' @seealso  Uses \link{seqEqual} for testing equivalence between pairs.
 //'           See \link{pairwiseDist} for generating a sequence distance matrix.
-//'           
+//'
 //' @examples
 //' # Gaps and Ns will match any character
 //' seq <- c(A="ATGGC", B="ATGGG", C="ATGGG", D="AT--C", E="NTGGG")
 //' d <- pairwiseEqual(seq)
 //' rownames(d) <- colnames(d) <- seq
 //' d
-//' 
+//'
+//' # Ignore only Ns, so that gaps are not treated as matching any character
+//' d <- pairwiseEqual(seq, ignore="N")
+//' rownames(d) <- colnames(d) <- seq
+//' d
+//'
 //' @export
 // [[Rcpp::export]]
-LogicalMatrix pairwiseEqual(StringVector seq) {
+LogicalMatrix pairwiseEqual(StringVector seq,
+                            CharacterVector ignore=CharacterVector::create("N","-",".","?")) {
     
     // allocate the matrix we will return
     LogicalMatrix rmat(seq.length(), seq.length());
@@ -115,7 +123,7 @@ LogicalMatrix pairwiseEqual(StringVector seq) {
             std::string row_seq = as<std::string>(seq[i]);
             std::string col_seq = as<std::string>(seq[j]);
             
-            bool is_equal = seqEqual(row_seq, col_seq);
+            bool is_equal = seqEqual(row_seq, col_seq, ignore);
             
             // write to output matrix
             rmat(i,j) = is_equal;